# customer-master/openapi/customer-master-api.yaml
openapi: 3.1.0
info:
  title: Customer Master API
  version: "1"
servers:
- url: /
paths:
...
  /api/states:
    get:
      operationId: listStates
      parameters:
      - description: "Case-insensitive name fragment, at most 10 characters"
        in: query
        name: nameContains
        required: false
        schema:
          type: string
          maxLength: 10
      - description: "Order of the list: by name (default) or by code"
        in: query
        name: sort
        required: false
        schema:
          type: string
          default: name
          enum:
          - name
          - code
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StateResponse"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: APP0400
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: APP0401
        "405":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: APP0400 (method not allowed; Allow lists the supported methods)
        "406":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: APP0400 (not acceptable)
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: DEM9999 (unexpected failure; errorId matches the server's ERROR
            log line)
      security:
      - basicAuth: []
      summary: "USA states, optionally filtered by name"
      tags:
      - States
components:
  schemas:
...
    CustomerResponse:
      type: object
      properties:
        acctMgr:
          type: string
        acctPhone:
          type: string
        active:
          type: string
        addr:
          type: string
        chgTime:
          type:
          - string
          - "null"
          format: date-time
        chgUser:
          type: string
        city:
          type: string
        corpPhone:
          type: string
        custId:
          type: string
        name:
          type: string
        state:
          type: string
        version:
          type: integer
          format: int64
        zip:
          type: string
      required:
      - acctMgr
      - acctPhone
      - active
      - addr
      - chgTime
      - chgUser
      - city
      - corpPhone
      - custId
      - name
      - state
      - version
      - zip
...
    FieldError:
      type: object
      description: One field at fault in a problem's errors
      properties:
        code:
          type: string
          description: The catalog key
          example: DEM0502
        field:
          type: string
          description: "The request body's JSON property (name, addr, city, state,\
            \ zip, corpPhone, acctMgr, acctPhone, active; version on update; purpose\
            \ on review) or the path or query parameter (custId; name, city, state,\
            \ includeInactive, size, cursor on search; nameContains, sort on the state\
            \ list) at fault"
          example: name
        message:
          type: string
          description: The catalog text with the arguments substituted
          example: "Name: Must not be blank"
      required:
      - code
      - field
      - message
...
    Problem:
      type: object
      description: "RFC 9457 problem: the body of every error response. type, title,\
        \ status, detail, code and args are always present; the other members only\
        \ in the cases their descriptions name."
      properties:
        args:
          type: array
          description: "The values substituted into the catalog text, as strings;\
            \ empty when there are none"
          items:
            type: string
        code:
          type: string
          description: "The catalog key, for example APP0400, DEM0502 or DEM1002"
          example: DEM1002
        current:
          $ref: "#/components/schemas/CustomerResponse"
          description: "409 DEM1002 only: the customer as now stored, including its\
            \ version"
        detail:
          type: string
          description: The catalog text of code with args substituted
          example: Someone else changed record. Review data.
        errorId:
          type: string
          format: uuid
          description: "500 DEM9999 only: the correlation id, also written to the\
            \ server's ERROR log line"
        errors:
          type: array
          description: "The fields at fault; the first entry receives focus. Present\
            \ when the failure names a request property or parameter: 400 APP0400\
            \ on a body property or a path or query parameter, 400 DEM0007 on state,\
            \ 422 DEM0501, DEM0502 or DEM0503 on the field whose rule failed, and\
            \ 422 DEM9898 on addr, city, state and zip, in that order. Absent when\
            \ it names none, as for a malformed body, an unknown property, content\
            \ after the JSON value, an unknown route, 405, 406, 415, 401, 403, 404\
            \ DEM0599, 409, 502, 503 and 500"
          items:
            $ref: "#/components/schemas/FieldError"
          minItems: 1
        instance:
          type: string
          format: uri-reference
          description: The request path; absent only when the path is not a usable
            URI reference
          example: /api/customers/EEEF
        stateAccepted:
          type: string
          description: "Failures of POST /api/customers/review only (422 and 502):\
            \ the normalized State, present when the State rule passed before the\
            \ failure and absent when the failure is at the State rule or before it"
          example: ME
        status:
          type: integer
          format: int32
          description: The HTTP status of the response
          example: 409
        title:
          type: string
          description: The HTTP reason phrase of status
          example: Conflict
        type:
          type: string
          format: uri
          description: "urn:customer-master:problem: followed by the catalog key"
          example: urn:customer-master:problem:DEM1002
      required:
      - args
      - code
      - detail
      - status
      - title
      - type
...
    StateResponse:
      type: object
      properties:
        name:
          type: string
        state:
          type: string
      required:
      - name
      - state
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http